extract worker's brigadier polyfill into its own module
