+++ v2/brigadier-polyfill/src/index.ts
@@ -1,11 +1,11 @@
-// This module provides the worker with a drop-in replacement for the REAL
-// Brigadier module. This drop-in replacement seamlessly replaces certain
+// This module provides the worker with a polyfill (drop-in replacement) for the
+// REAL Brigadier module. This drop-in replacement seamlessly replaces certain
 // Brigadier behaviors with more sophisticated alternatives. For instance, the
 // no-op job.run() function is replaced with one that communicates with the
 // Brigade API server.
 
 // Export these things directly from Brigadier
-export { Container, ConcurrentGroup, Event, JobHost, SerialGroup }  from "../../brigadier/src"
+export { Container, ConcurrentGroup, Event, JobHost, SerialGroup }  from "@brigadecore/brigadier"
 
 // These are custom implementations of resources ordinarily found in Brigadier
 export { events } from "./events"
