+++ v2/brigadier-polyfill/src/logger.ts
